fix: flaky test due to test assert before file write is complete
