+++ tests/listCommand.test.js
@@ -4,7 +4,10 @@ const { EOL } = require("node:os")
 const { main } = require("../index.js")
 const { SAVE_FILE_PATH } = require("../storage/file.js")
 
-describe("add cli task", () => {
+const sleep = (ms) => new Promise((resolve) => setTimeout(resolve, ms))
+const WAIT_TIME_FOR_IO_IN_MS = 500
+
+describe("list command", () => {
   beforeEach(() => {
     deleteSaveFile()
   })
@@ -33,47 +36,58 @@ describe("add cli task", () => {
 
   it("should list zero tasks when zero tasks are present", async () => {
     const { stdin, stdout } = mockStdinAndStdout()
-    main()
+    async function onExit() {
+      await sleep(WAIT_TIME_FOR_IO_IN_MS)
+      const saveFileData = readSaveFileContent()
+      expect(saveFileData).toBe("", "Should have no saved tasks")
+      await expect(stdout.output).not.toMatch(
+        /Invalid Command \[list\] given/,
+        "Should not have invalid command message"
+      )
+      await expect(stdout.output).toMatch(/No Tasks Available/)
+    }
+    main(onExit)
     stdin.send(`list` + EOL)
     stdin.end()
     stdout.stop()
-    const saveFileData = readSaveFileContent()
-    expect(saveFileData).toBe("", "Should have no saved tasks")
-    await expect(stdout.output).not.toMatch(
-      /Invalid Command \[list\] given/,
-      "Should not have invalid command message"
-    )
-    await expect(stdout.output).toMatch(/No Tasks Available/)
   })
 
   it("should reject list command with invalid status keyword", async () => {
     const invalidListCommand = `list special`
     const { stdin, stdout } = mockStdinAndStdout()
-    main()
+    async function onExit() {
+      await sleep(WAIT_TIME_FOR_IO_IN_MS)
+      await expect(stdout.output).toMatch(
+        new RegExp(
+          String.raw`Invalid Command \[${invalidListCommand}\] given.`
+        ),
+        "Invalid Command message should be displayed to user"
+      )
+    }
+    main(onExit)
     stdin.send(invalidListCommand + EOL)
     stdin.end()
     stdout.stop()
-    await expect(stdout.output).toMatch(
-      new RegExp(String.raw`Invalid Command \[${invalidListCommand}\] given.`),
-      "Invalid Command message should be displayed to user"
-    )
   })
 
   it("should list two tasks after adding two tasks", async () => {
     const { stdin, stdout } = mockStdinAndStdout()
-    main()
+    async function onExit() {
+      await sleep(WAIT_TIME_FOR_IO_IN_MS)
+      await expect(stdout.output).toMatch(
+        /1\) first task \[todo\]/,
+        "Should list first task"
+      )
+      await expect(stdout.output).toMatch(
+        /2\) second task \[todo\]/,
+        "Should list second task"
+      )
+    }
+    main(onExit)
     stdin.send(`add "first task"` + EOL)
     stdin.send(`add "second task"` + EOL)
     stdin.send(`list ` + EOL)
     stdin.end()
     stdout.stop()
-    await expect(stdout.output).toMatch(
-      /1\) first task \[todo\]/,
-      "Should list first task"
-    )
-    await expect(stdout.output).toMatch(
-      /2\) second task \[todo\]/,
-      "Should list second task"
-    )
   })
 })
